slamd-120: Moved run_experiment() tests to new file
